!!! CHANGED SERIALIZATION FORMAT NOW: Header version: half-byte Version history length: half-byte Version: byte 1*N {Version hash: short} *N {FIELD: BINARY}
@FieldVersion -> @SinceVersion
Annotations from fields moved to properties
Migrated test system to JUnit
Refactorings and performance improvements

+++ icegem-core/src/main/java/com/googlecode/icegem/serialization/BeanVersion.java
@@ -11,5 +11,5 @@ import java.lang.annotation.Target;
 @Target(ElementType.TYPE)
 @Retention(RetentionPolicy.RUNTIME)
 public @interface BeanVersion {
-    int value();
+    byte value();
 }
